i386-sse-4.c: Use 'float' as the base type for FP types.
	* gcc.dg/i386-sse-4.c: Use 'float' as the base type for FP types.
	* gcc.dg/i386-sse-5.c: Likewise.

From-SVN: r89138

+++ gcc/testsuite/gcc.dg/i386-sse-5.c
@@ -1,6 +1,6 @@
 /* { dg-do compile { target i?86-*-* } } */
 /* { dg-options "-Winline -O2 -march=i386" } */
-typedef int v2df __attribute__ ((mode(V2DF)));
+typedef float v2df __attribute__ ((mode(V2DF)));
 v2df p;
 q(v2df t)
 {			 /* { dg-warning "SSE" "" } */
